feat(15): added conditional automatic scrolling
Whenever the user scrolls up, instead of automatically scrolling down on a new message event, a small button appears which, if clicked, scrolls the chatbox messages list downwards.
Also, code review suggestions have been considered, however, the start phases and dummy data will be kept on this branch until we merge it with the ETS, otherwise the dummy room can't be accessed.