feat(17): moved emoji pop-up handlers into a hook
Instead of having a global window.onclick event handler which would raise a lot of errors on pages other than single-room-page, an OpenEmojiPopup hook has been created. This hook adds event handlers to both the button and the pop-up and there is a new feature that improves the UX: the pop-up only closes when hovering off/pointer leaves the pop-up window.